starboard: Use detailed SB_CHECK logging for all builds (#7174)
In an "official" build (when `SB_LOGGING_IS_OFFICIAL_BUILD` is 1),
`SB_CHECK(condition)` expands to a direct call to
`::starboard::Break()`, providing no diagnostic information in the
resulting crash.

By forcing the non-official path, a failing `SB_CHECK` will instead go
through the `SB_LOG(FATAL)` macro. This generates a rich log message
containing the failed condition and a stack trace, which is invaluable
for debugging.

While this introduces overhead, it is incurred only upon failure and,
critically, occurs *after* the condition is checked. This ensures that
the timing of the check is not altered, preventing this change from
masking or creating race conditions.

With this change,  Cobalt.apk gold build size increases by 61KB.
- BEFORE:  79,610 KB
- AFTER:     79,549 KB

Bug: 445154341