Move video geometry setter code out of OverlayStrategyUnderlayStarboard
`OverlayStrategyUnderlayStarboard` is not used by tvOS, as the code path in
Viz for Apple platforms is different and none of the
OverlayProcessorUsingStrategy classes are used.

As a first step towards making hole punching work for tvOS, move the video
geometry setter out of `OverlayStrategyUnderlayStarboard` and into a
separate file that can be used by both the existing class as well as the
upcoming tvOS code in Viz.

Bug: 459366150
Bug: 456692793