tvOS: Move the UIView that holds SbPlayer-rendered videos to Cobalt Shell
This is the Starboard/Cobalt Shell side of #8319.

The existing code imported from C25 creates a player container UIView in
`application_view.mm` that was used to hold the video renderer UIViews
created in `av_sample_buffer_video_renderer.mm` that SbPlayer would render
to. This no longer works in main, as it is Cobalt Shell that is responsible
for creating the UI that displays web contents and everything else now.

This commit makes `ContentWindowShellDelegate` in Cobalt Shell responsible
for creating a player container UIView that will hold the videos, and
integration with Starboard is made via the main `SBDStarboardApplication`
protocol which now has two methods for registering the container view as
well as for adding views to it: the former is called by Cobalt Shell, the
latter is called by `av_sample_buffer_video_renderer.mm` on the Starboard
side.

Note: this just sets up the right UIView hierarchy in the right
locations. For the actual video rendering to work, we need to make the
Apple code path in Viz work with video hole quads (#8319).

Bug: 459366150