[tvOS] Make the Apple code path in Viz work with video hole quads
This is part of the required changes to make SBPlayer-rendered videos show
on the screen together with the rest of the web contents.

The idea here is to make hole punching to work by having the videos handled
by SbPlayer be shown as underlays: the UIViews with videos are added before
the web contents view and are thus supposed to be shown "under" it (this
will be done in a separate PR after #8116 lands).

On the Viz side, this is done by treating video hole quads (which are
already added by Starboard media code in `//media`) specially: when found,
these quads disable the default quads-to-CALayers code path that normally
renders each quad as a separate CALayer (with other CALayers underneath,
managed and described by `CARendererLayerTree` and [1]).

[1] https://docs.google.com/document/d/1DtSN9zzvCF44_FQPM7ie01UxGHagQ66zfF5L9HnigQY/edit?usp=sharing

By doing so:

- `CALayerOverlayProcessor::PutForcedOverlayContentIntoUnderlays()`
  and `CALayerOverlayProcessor::PutQuadInSeparateOverlay()` are always
  called, and we use `PutQuadInSeparateOverlay()` to replace the video
  hole quad with a transparent quad just like the existing code in
  `OverlayStrategyUnderlayStarboard`.
- The entire page is received by `CARendererLayerTree` as a single
  CALayer (like in C25), which includes the transparent quad created above,
  effectively making the video UIViews poke through the web contents.

Bug: 459366150