starboard: Flatten `starboard::shared` namespace used for linux build (#7159)
[go/cobalt-flatten-starboard-namespace](https://goto.google.com/cobalt-flatten-starboard-namespace)

Currently, Linux-specific code uses the `starboard::shared` namespace
instead of `starboard::linux`. This workaround exists because `linux`
often conflicts with predefined macros or keywords in some build
toolchains.

This situation is misleading, as it results in Linux-specific methods
residing within the generic shared namespace. Flattening the `starboard`
namespace, which this PR contributes to, resolves this confusion by
removing the ambiguous nested namespace.

Bug: 441955897