starboard: Move `DecodedAudio` and `Buffer` into flattened namespace (#7197)
[go/cobalt-flatten-starboard-namespace](http.go/cobalt-flatten-starboard-namespace)

This change moves the `DecodedAudio` and `Buffer` classes from the
`starboard::shared::starboard::player` namespace into the top-level
`starboard` namespace.

This refactoring simplifies the code by removing a layer of namespace
nesting and eliminating the need for local `typedefs` in various audio
decoder and renderer components. This makes the code cleaner and more
consistent with the ongoing namespace flattening effort.

Bug: 441955897