cobalt/android: Adding WebContents Listener to Shell.java (#7208)
Due to async processing, there's a small chance that webContents is not
finished being created by the time `initializeJavaBridge()` and
`getStarboardBridge.setWebContents(getActiveWebContents())` are called
in the initialization process. This PR adds a listener
`OnWebcontentsReadyListener` passed through the ShellManger to the Shell
and only proceeds with finishing the initialization process once
webContents is initialized and available.

Bug: 439066169