starboard: Move PlayerComponents into flattened namespace
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the `PlayerComponents` class and its factory from the `starboard::shared::starboard::player::filter` namespace into the top-level `starboard` namespace.

This refactoring is part of the ongoing effort to simplify the Starboard namespace structure. It removes unnecessary nesting and makes the player utilities more accessible and consistent with the flattened namespace design.

Bug: 441955897