starboard: Move player classes into flattened namespace
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the core player and filter-based player classes from the `starboard::shared::starboard::player` and `starboard::shared::starboard::player::filter` namespaces into the top-level `starboard` namespace.

This refactoring is part of the ongoing effort to simplify the Starboard
namespace structure. It removes unnecessary nesting and makes the player implementation more accessible and consistent with the flattened namespace design.

Bug: 441955897