starboard: Flatten signal handling namespace (#7226)
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the signal handling functions from the
`starboard::shared::signal` namespace into the top-level `starboard`
namespace.

This refactoring is part of the ongoing effort to simplify the Starboard
namespace structure. It removes unnecessary nesting and makes the signal
handling utilities more accessible and consistent with the flattened
namespace design.

Bug: 441955897