starboard: Move drm test helper function to flattened namespace (#7227)
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the drm test helper functions and constants from the
`starboard::shared::starboard::drm` namespace into the top-level
`starboard` namespace.

This refactoring is part of the ongoing effort to simplify the Starboard
namespace structure. It removes unnecessary nesting and makes the drm
test utilities more accessible and consistent with the flattened
namespace design.

Bug: 441955897