starboard: Flatten X11 and dev_input namespaces
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the X11 and dev_input classes from the
`starboard::shared::x11` and `starboard::shared::dev_input` namespaces
into the top-level `starboard` namespace.

This refactoring is part of the ongoing effort to simplify the Starboard
namespace structure. It removes unnecessary nesting and makes the X11 and
dev_input utilities more accessible and consistent with the flattened
namespace design.

Bug: 441955897