starboard: Flatten elf_loader and crashpad namespaces
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the elf_loader and crashpad classes from the
`starboard::elf_loader` and `third_party::crashpad` namespaces
into the top-level `starboard` and `crashpad` namespaces respectively.

This refactoring is part of the ongoing effort to simplify the Starboard
namespace structure. It removes unnecessary nesting and makes the elf_loader
and crashpad utilities more accessible and consistent with the flattened
namespace design.

Bug: 441955897