starboard: Flatten loader_app namespace
[go/cobalt-flatten-starboard-namespace](http://go/cobalt-flatten-starboard-namespace)

This change moves the loader_app classes from the
`starboard::loader_app` namespace into the top-level `loader_app`
namespace.

This refactoring is part of the ongoing effort to simplify the Starboard
namespace structure. It removes unnecessary nesting and makes the loader_app
utilities more accessible and consistent with the flattened namespace design.

Bug: 441955897