[tvOS] Do not set RPATH in starboard GN config (#7230)
Basically a reland of the most conservative version of #7179 that does
not remove on all platforms, as that still needs to be worked out in a
way that does not break linux-x64x11 at runtime at least.

The syntax is not recognized by LLD, leading to errors like

    ld64.lld: error: unknown argument '-rpath=/foo/bar'

when linking executables both both host and target.

Since everything works fine without setting any RPATH, just avoid trying
to set it altogether.

Bug: 432642933