media: Add BUFFER_FLAG_DECODE_ONLY Experiment
This PR adds an experiment to allow the use of the MediaCodec flag
BUFFER_FLAG_DECODE_ONLY on non-tunneled playbacks. This flag is
originally only used for tunneled playbacks, but it is possible that
non-tunneled playback performance can be improved with this flag as
well. This experiment is guarded by a starboard feature.

Bug: 423927047