cobalt: Introduce tvOS user_agent and //base logic (#7146)
Follow up to #7076. 

This CL introduces `InitializePlatformDependentFieldsTvOS()` to
`UserAgentPlatformInfo` and populates the known bits inside to
generate some common parts. set_rasterizer_type() is moved to
the platform-specific methods, and takes the chance to gather
starboard::GetSystemPropertyString() calls together and under
the existing `avoid_access_to_starboard_for_testing_` umbrella.

It also adds the base::SbSysInfo::* functions needed and 
enables the unittest for that platform (note that 
`BUILDFLAG(IS_IOS)` should be replaced with 
`BUILDFLAG(IS_IOS_TVOS` after the rebase, a TODO is left).



Bug: 442760345
Bug: 444510191