POSIX: Add setpriority and getpriority to Starboard (#7096)
This PR adds the process function `setpriority` and `getpriority` as
Starboard Wrappers around the platform
implementations. While `setpriority` is the only "leaking" symbol,
`getpriority` was chosen to be supported as well as it can help us test
the functionality of `setpriority`. Both symbols are wrapped due to
there being special macros that determine what kind of process is being
called.

The unit tests for this implementation were constructed with help from
Gemini, and then edited for correctness.

Bug: 412648662