Enable basic crash reporting for evergreen-x64 (#7280)
This enables Crashpad components to be built and to snapshot+upload
crashes for evergreen-x64.

This implementation does not yet have feature parity with c25; most of
the significant features and customizations yet to be implemented are
tracked by child issues under b/406511608.

The remaining open questions from go/crashpad-design-questions would not
have a major impact on this implementation and so should not prevent us
from merging this PR.

Significant changes vs. c25:
* The Crashpad client and handler are built on top of //base rather than
mini_chromium. This follows the approach used in Chromium.
* The Crashpad handler is installed from the elf_loader_sandbox and
loader_app. This obviates the need for partners to do the installation
in their ports and also removes a //starboard -> //base -> //starboard
dependency cycle.

Issue: 406511608