Set STARBOARD_GL_TYPE_* in //starboard/build/config:starboard
For some reason, the egl_and_gles:defines config was not being propagated on
linux-x64x11-modular and evergreen-x64. The only reason I can think of is a
problem caused by `//starboard:starboard` being defined with a different
toolchain.

Use a big hammer and just move the logic to
`//starboard/build/config:starboard`, which is depended upon in multiple
targets anyway.