POSIX: Add getrlimit to Starboard (#7351)
This PR adds the process function getrlimit as a registered Starboard
API symbol. This function is implemented with a wrapper function as it
takes in a non-standard struct value, the rlimit struct.

The unit tests for this implementation were constructed with the help
from Gemini, and then edited for correctness.

Bug: 412648662