android: Disabling esa commandLineArgs for gold builds (#7404)
As per
[go/chrobalt-system-dev-dd](https://goto.google.com/chrobalt-system-dev-dd),
devtools access should not be allowed in release builds. Adding a new
versionInfo isReleaseBuild to check for release build. Couldn't use the
existing isReleaseBuild() starboard function as it's not ready by the
time createContent() is called. Checking in both createContent() and
getArgs() disables esa commandLineArgs for release builds.

Bug: 446019286