FakeGraphicsContextProvider: Clean up display initialization code in InitializeEGL() (#7347)
The logic added in #5468, #5950 and #6438 was not working as expected:

* This file does not include the EGL headers directly, so nothing ever sets
  `EGL_VERSION_1_5` and the EGL 1.4 code path was always being used.
* Use of `eglGetPlatformDisplay()` with ANGLE parameters only makes use
when
  ANGLE is used, which at the moment is the case only for linux-x64x11.
  On platforms that do not use ANGLE but do provide
  `eglGetPlatformDisplayEXT()`, the call was failing because ANGLE
  attributes were being used.

The new logic is simpler and should work with Android, linux-x64x11,
Raspi2, RDK and other platforms:

* Remove all checks for `EGL_VERSION_1_5`.
* Only call `eglGetPlatformDisplay()` when ANGLE is in use, as all the
platform being passed and its attributes are ANGLE-specific at the
moment.
* Call `eglGetDisplay()` in all other cases (i.e. Android, Raspi2, RDK
etc).
  This was already the case for Android; Raspi2 does not have
`eglGetPlatformDisplayEXT()` so it was always doing this too; RDK has a
custom `SbEglGetDisplay()` that tries to use `eglGetPlatformDisplayEXT()`
when it is available (without the wrong ANGLE parameters that were being
passed here).

Note that other platforms that do not use ANGLE but have EGL 1.5 could
also call `eglGetPlatformDisplay()`, but they need to pass the right
attributes and a valid platform value to the function.

Also note that at the moment raspi2 and RDK skip all NPLB tests using
this class.

---

This PR contains 2 commits but since they are probably going to be
squashed, some explanation about the changes to `starboard/egl_and_gles`
are in order:

* Clean up list of valid `gl_type` values in the documentation as well
as in the validation (i.e. "glimp" and "none" were either not documented or
not allowed).
* Define the `egl_and_gles_angle` group, there is no need to define it
  conditionally.
* Set the `STARBOARD_GL_TYPE_{ANGLE,SYSTEM_GLES2}` macros depending on
  `gl_type`'s value. This is useful when determining whether we can use
  ANGLE values (in FakeGraphicsContextProvider, for example).

---

Bug: 402459390
Bug: 420988131
Bug: 432259527
Fixed: 447076985
Fixed: 447478802

original-hexsha: 451f80ac455cbef3a68ddbe8b656553ae0e44869