media/starboard: make SbPlayerSetBoundsHelper not ref-counted (#7092)
This CL/PR makes SbPlayerSetBoundsHelper not-ref counted: Its only 
callsite is StarboardRenderer::OnVideoGeometryChange() which IIUC 
is a Mojo callback; so it could be on a Task Runner different
from |task_runner_|: left a CHECK and if it trips, the real way to 
avoid a problem is via PostTask.

This CL also removes unnecessary expansion in
SbPlayerSetBoundsHelper::SetBounds
parameter(s), using gfx::Rect directly.

Fixed: 448749860