starboard: Make SbSystemRequestStop a no-op on Android (#7470)
`SbSystemRequestStop()` was implemented using a legacy JNI call, but
this function is not used.

This change removes the unnecessary JNI implementation and replaces it
with a warning log. This simplifies the native codebase by removing an
unused JNI call.

Bug: 450024477

original-hexsha: 7243a636a2ca6ce8ea15c194e86660d03b641c31