POSIX: Add sched_getaffinity to Starboard (#7439)
This PR adds the Linux function `sched_getaffinity` to Starboard with a
wrapper function. This symbol is wrapped due to it taking in a
non-standard defined struct, `cpu_set_t`.

The unit tests for this implementation were constructed with help from
Gemini, and then edited for correctness.

Bug: 412648662

original-hexsha: 1060ea0c27b910fa3e39af286982833d16bda472