common: Remove codec libraries pulled from //media (#7053)
Top level //media transitively depend on many media decoding/encoding
libraries, which are either not needed in Cobalt, or pulled from 
//starboard dependencies. This CL removes all/many of those
dependencies.

**Instruction for rebasers** 
Part of this PR landed as an upstream CL, concretely around 143.0.7460.0.
It could not be cleanly backported, so once Cobalt passes said Chrome
version, the modifications in this PR for the files:
- `//media/base/supported_types_unittest.cc`
- `//media/cdm/library_cdm/clear_key_cdm/cdm_video_decoder.cc`
can simply be dropped.

Bug: 421171908 
Bug: 437151722