Fix the touch panel crash (#7214)
A crash occurs during vertical scrolling. The crash was caused by the
scroll engine receiving a null pointer.

The `FindFirstElementWithScrollType` function in
`topmost_event_target.cc` was returning a scrollable element that did
not have an associated `UiNavItem` for vertical scroll events. This led
to a null pointer being passed to the `scroll_engine`, causing the
application to crash.

This change ensure scrollable element has a valid `UiNavItem` before
calling scroll_engine

Bug: 426634950