Remove customization to send EG info addr via IPC
For Evergreen, the crashpad client needs to send the handler the address
to a struct, in the client process, containing information about the
Evergreen shared library. In the main branch we now only start the
handler after a crash and in this case the client sends that address to
the handler in its argv. We can remove the customization that allows the
client, when the handler is started at app start time, to send the
address to the already running handler process using IPC.

The conditional inclusion of kTypeEvergreenInfo in the
ClientToServerMessage struct has been causing some compiler errors for
some build configurations. For the reasons provided above, it turns out
we can just remove it entirely.

Issue: 447430712
Issue: 406511608