tvos: AVSBVideoRenderer: Use onApplicationMainThread instead of dispatch_async()
Some PlayerComponentsTests that use test data that is _not_ audio-only were
timing out and being flaky because the blocks in `dispatch_async()` were
either never being run or took too long to run.

This happens because, contrary to the Cobalt Shell code, all this code in
the class is running on the main thread together with the unit tests
themselves, so the main thread ends up being blocked.

The tests themselves should probably run this code in a different thread to
better reflect what production code does (I have filed b/450473125 to track
this), but for now what we can do is just run those blocks synchronously
when the code is being run from the main thread. This is exactly what
`onApplicationMainThread()` does: when it is invoked from a different
thread, it schedules the block to run later with `dispatch_async`; when it
is run from the main thread, it just invokes the block immediately.

Bug: 459366150
Bug: 456692793