android: Remove unused legacy JNI for audio permissions (#7509)
Removes the native implementation of `nativeHandlePermission` from
`microphone_impl.cc`. This JNI function is no longer called from the
Java side after the audio permission flow was refactored in #5055.

This change is part of the larger effort to migrate from legacy JNI to
the modern JNI generator. As part of this migration, it is best practice
to remove unused legacy JNI methods to clean up the codebase and avoid
confusion.

Bug: 450024477

original-hexsha: 4dee3daa525b239ccb40438091f34a5aca5ce26e