Add custom InsertAnnotationForHandler API from c25
This enables Crashpad annotations to be provided by a client after the
signal handler has been installed. This is needed to enable annotations
to be set by the Cobalt layer, including on behalf of a web app.

These changes are mostly just ported forward c25, with a couple
significant updates:
* Use of locks from //base/synchronization instead of Starboard
* The scope is much smaller since we are not currently supporting the
  RequestCrashDumpHandler SignalHandler in Chrobalt.

A separate PR will wire up the CrashHandler Starboard extension so that
this API is used by the setString() web API for 3P.

Issue: 446908034