Improve locking.
 * Make the file locks and the open file lock (ofl) recursive to match
   the futex based behavior.
 * Don't use musl's empty lseek, just map __lseek to seek with a macro.
 * Add some more missing syscall wrapper macros.
 * Initialize the file mutex right away in fdopen.
 * Use the efficient syscall() to libc wrapper also outside of musl.

Note: File locks are not actually necessary if the FILE* is never used
from another thread. Musl handles this by only requiring locking on the
files that are open when a thread is created. That has one gap though,
because it does not take into account when FILE* are passed between
threads. For now, we always use file locks with Starboard to ensure
correct locking in all cases.