linux: Remove forced GPU memory limit for 3p platforms (#7558)
Removing the kForceGpuMemAvailableMb switch for third-party platforms.
This setting was causing significant rendering issues, including
flickering and occasional black screens.

This change was previously reverted for the Android platform due to the
same instability.

While forcing the memory size to 64MB resolved the flickering, memory
measurements showed that the total RSS was the same as when the value is
not set at all. To avoid hardcoding a value that the compositor can
determine more effectively, this switch is being removed.

The original PR for adding this:
https://github.com/youtube/cobalt/pull/5945/

Bug: 452419910

original-hexsha: bf52231c1a649cd894096f1f9388882d645cb8c1