Implement CrashAnnotator interface for 3P (#7548)
This implements the `CrashLog.setString` web API on 3P platforms
(confirmed with local testing).

This change involves introducing a new version of the CrashHandler
Starboard Extension and wiring it up for 3P platforms. The new version
enables a callback to be provided to the extension implementation so
that it can delegate implementation of the `SetString` function to a
dependency. This allows us to avoid a build dependency cycle for 3P
platforms: `//starboard/crashpad_wrapper` indirectly depends on
`//starboard:starboard_group` via `//base`, so we want to avoid also
having `//starboard:starboard_group` depend on
`//starboard/crashpad_wrapper`. We wouldn't have had this problem in c25
because the Crashpad client library was built on top of `mini_chromium`
instead of `//base`.

Issue: 446908034
Issue: 400709539